refactor: replace unnecessary useEffect w/ useMemo
refactor k8ObjectMap to k8ObjectOptionType conversion to Utils.tsx
refactor out unnecessary state & useEffect call
