+++ .eslintignore
@@ -146,7 +146,6 @@ src/components/ResourceBrowser/ResourceList/ResourceFilterOptions.tsx
 src/components/ResourceBrowser/ResourceList/ResourceList.component.tsx
 src/components/ResourceBrowser/ResourceList/ResourceList.tsx
 src/components/ResourceBrowser/ResourceList/ResourceListEmptyState.tsx
-src/components/ResourceBrowser/ResourceList/Sidebar.tsx
 src/components/ResourceBrowser/ResourceList/__mocks__/Sidebar.mock.ts
 src/components/ResourceBrowser/Types.ts
 src/components/ResourceBrowser/Utils.ts
