+++ src/components/ResourceBrowser/ResourceList/Sidebar.tsx
@@ -1,7 +1,7 @@
-import React, { Fragment, useEffect, useRef, useState } from 'react'
+import React, { Fragment, useEffect, useRef, useState, useMemo, useCallback } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
-import ReactSelect, { GroupBase, InputActionMeta } from 'react-select'
-import Select, { FormatOptionLabelMeta } from 'react-select/dist/declarations/src/Select'
+import ReactSelect, { InputActionMeta, GroupBase } from 'react-select'
+import Select, { FormatOptionLabelMeta } from 'react-select/base'
 import { withShortcut, IWithShortcut } from 'react-keybind'
 import { useAsync } from '@devtron-labs/devtron-fe-common-lib'
 import { URLS } from '../../../config'
@@ -18,6 +18,7 @@ import { K8S_EMPTY_GROUP, KIND_SEARCH_COMMON_STYLES, SIDEBAR_KEYS } from '../Con
 import { KindSearchClearIndicator, KindSearchValueContainer } from './ResourceList.component'
 import { getK8Abbreviates } from '../ResourceBrowser.service'
 import { swap } from '../../common/helpers/util'
+import { convertK8sObjectMapToOptionsList } from '../Utils'
 
 const Sidebar = ({
     k8SObjectMap,
@@ -37,11 +38,20 @@ const Sidebar = ({
         group: string
     }>()
     const [searchText, setSearchText] = useState('')
-    const [k8sObjectOptionsList, setK8sObjectOptionsList] = useState<K8sObjectOptionType[]>([])
-    const [k8AbbreviatesLoading, k8Abbreviates] = useAsync(getK8Abbreviates)
+    const [, k8Abbreviates] = useAsync(getK8Abbreviates)
     const sideBarElementRef = useRef<HTMLDivElement>(null)
     const preventScrollRef = useRef<boolean>(false)
     const searchInputRef = useRef<Select<K8sObjectOptionType, false, GroupBase<K8sObjectOptionType>>>(null)
+    const k8sObjectOptionsList = useMemo(() => convertK8sObjectMapToOptionsList(k8SObjectMap), [k8SObjectMap?.size])
+
+    const handleInputShortcut = (e: React.KeyboardEvent<HTMLDivElement>) => {
+        const _key = e.key
+        if (_key === 'k') {
+            searchInputRef.current?.focus()
+        } else if (_key === 'Escape' || _key === 'Esc') {
+            searchInputRef.current?.blur()
+        }
+    }
 
     useEffect(() => {
         if (!isCreateModalOpen) {
@@ -58,99 +68,14 @@ const Sidebar = ({
             if (!preventScrollRef.current && sideBarElementRef.current) {
                 sideBarElementRef.current.scrollIntoView({ block: 'center' })
             }
-
-            if (!k8sObjectOptionsList.length) {
-                covertK8sMapToOptionsList()
-            }
         }
-    }, [k8SObjectMap?.size, sideBarElementRef.current])
+    }, [sideBarElementRef.current])
 
-    const handleInputShortcut = (e: React.KeyboardEvent<any>) => {
-        const _key = e.key
-        if (_key === 'k') {
-            searchInputRef.current?.focus()
-        } else if (_key === 'Escape' || _key === 'Esc') {
-            searchInputRef.current?.blur()
-        }
-    }
-
-    const covertK8sMapToOptionsList = () => {
-        let isNamespacesAvailable
-        let isEventsAvailable
-        const _k8sObjectOptionsList = [...k8SObjectMap.values()].flatMap((k8sObject) => {
-            return [...k8sObject.child.entries()].flatMap(([key, value]) => {
-                const keyLowerCased = key.toLowerCase()
-                if (
-                    keyLowerCased === 'node' ||
-                    keyLowerCased === SIDEBAR_KEYS.namespaceGVK.Kind.toLowerCase() ||
-                    keyLowerCased === SIDEBAR_KEYS.eventGVK.Kind.toLowerCase()
-                ) {
-                    isNamespacesAvailable =
-                        isNamespacesAvailable || keyLowerCased === SIDEBAR_KEYS.namespaceGVK.Kind.toLowerCase()
-                    isEventsAvailable = isEventsAvailable || keyLowerCased === SIDEBAR_KEYS.eventGVK.Kind.toLowerCase()
-                    return []
-                }
-
-                return value.data.map((childData) => {
-                    return {
-                        label: childData.gvk.Kind,
-                        value: childData.gvk.Group || K8S_EMPTY_GROUP,
-                        dataset: {
-                            group: childData.gvk.Group,
-                            version: childData.gvk.Version,
-                            kind: childData.gvk.Kind,
-                            namespaced: `${childData.namespaced}`,
-                            grouped: `${k8sObject.child.size > 1}`,
-                        },
-                        groupName: value.data.length === 1 ? k8sObject.name : `${k8sObject.name}/${key}`,
-                    }
-                })
-            })
-        })
-        if (isEventsAvailable) {
-            _k8sObjectOptionsList.push({
-                label: SIDEBAR_KEYS.events as Nodes,
-                value: K8S_EMPTY_GROUP,
-                dataset: {
-                    group: SIDEBAR_KEYS.eventGVK.Group,
-                    version: SIDEBAR_KEYS.eventGVK.Version,
-                    kind: SIDEBAR_KEYS.eventGVK.Kind as Nodes,
-                    namespaced: 'true',
-                    grouped: 'false',
-                },
-                groupName: '',
-            })
-        }
-        if (isNamespacesAvailable) {
-            _k8sObjectOptionsList.push({
-                label: SIDEBAR_KEYS.namespaces as Nodes,
-                value: K8S_EMPTY_GROUP,
-                dataset: {
-                    group: SIDEBAR_KEYS.namespaceGVK.Group,
-                    version: SIDEBAR_KEYS.namespaceGVK.Version,
-                    kind: SIDEBAR_KEYS.namespaceGVK.Kind as Nodes,
-                    namespaced: 'false',
-                    grouped: 'false',
-                },
-                groupName: '',
-            })
-        }
-        _k8sObjectOptionsList.push({
-            label: SIDEBAR_KEYS.nodes as Nodes,
-            value: K8S_EMPTY_GROUP,
-            dataset: {
-                group: SIDEBAR_KEYS.nodeGVK.Group,
-                version: SIDEBAR_KEYS.nodeGVK.Version,
-                kind: SIDEBAR_KEYS.nodeGVK.Kind,
-                namespaced: 'false',
-                grouped: 'false',
-            },
-            groupName: '',
-        })
-        setK8sObjectOptionsList(_k8sObjectOptionsList)
-    }
-
-    const selectNode = (e: any, groupName?: string, preventScroll?: boolean): void => {
+    const selectNode = (
+        e: React.MouseEvent<HTMLDivElement> | { currentTarget: Pick<K8sObjectOptionType, 'dataset'> },
+        groupName?: string,
+        preventScroll?: boolean,
+    ): void => {
         const _selectedKind = e.currentTarget.dataset.kind.toLowerCase()
         const _selectedGroup = e.currentTarget.dataset.group.toLowerCase()
 
@@ -164,7 +89,7 @@ const Sidebar = ({
             gvk: {
                 Group: e.currentTarget.dataset.group,
                 Version: e.currentTarget.dataset.version,
-                Kind: e.currentTarget.dataset.kind,
+                Kind: e.currentTarget.dataset.kind as Nodes,
             },
             isGrouped: e.currentTarget.dataset.grouped === 'true',
         }
@@ -251,7 +176,7 @@ const Sidebar = ({
                     <DropDown
                         className={`${value.isExpanded ? 'fcn-9' : 'fcn-5'}  rotate icon-dim-24 pointer`}
                         style={{
-                            ['--rotateBy' as any]: value.isExpanded ? '0deg' : '-90deg',
+                            ['--rotateBy' as string]: value.isExpanded ? '0deg' : '-90deg',
                         }}
                     />
                     <span className={`fs-14 ${value.isExpanded ? 'fw-6' : 'fw-4'} pointer w-100 pt-6 pb-6`}>{key}</span>
@@ -262,29 +187,29 @@ const Sidebar = ({
     }
 
     const handleInputChange = (newValue: string, actionMeta: InputActionMeta): void => {
+        // TODO: replace with enum after merge of feat/user-status-p3
         if (actionMeta.action !== 'input-change') {
             return
         }
 
         setSearchText(newValue)
+    }
 
-        if (!newValue || k8AbbreviatesLoading) {
-            return
+    const bringMatchedAbbreviatedOptionToFront = (options: K8sObjectOptionType[]): K8sObjectOptionType[] => {
+        if (!searchText || !k8Abbreviates) {
+            return options
         }
-
-        const abbreviate_expanded = k8Abbreviates?.[newValue.toLowerCase()]
-        if (!abbreviate_expanded) {
-            return
+        const lowerSearchText = searchText.toLowerCase()
+        if (!k8Abbreviates[lowerSearchText]) {
+            return options
         }
-
-        const loc = k8sObjectOptionsList.findIndex((value) => value.label.toLowerCase() === abbreviate_expanded)
-        if (loc === -1) {
-            return
+        const loc = k8sObjectOptionsList.findIndex(
+            (option) => k8Abbreviates[lowerSearchText] === option.label.toLowerCase(),
+        )
+        if (loc > -1) {
+            swap(options, loc, 0)
         }
-
-        swap(k8sObjectOptionsList, loc, 0);
-
-        setK8sObjectOptionsList([...k8sObjectOptionsList])
+        return options
     }
 
     const hideMenu = () => {
@@ -304,34 +229,38 @@ const Sidebar = ({
             option.groupName,
             option.label !== (SIDEBAR_KEYS.namespaces as Nodes) &&
                 option.label !== (SIDEBAR_KEYS.events as Nodes) &&
-                option.label !== (SIDEBAR_KEYS.nodes as Nodes)
+                option.label !== (SIDEBAR_KEYS.nodes as Nodes),
         )
     }
 
-    function formatOptionLabel(option: K8sObjectOptionType, formatOptionLabelMeta: FormatOptionLabelMeta<any>) {
-        return (
-            <div className="flex left column">
-                {!formatOptionLabelMeta.inputValue ? (
-                    <span className="w-100 dc__ellipsis-right">{option.label}</span>
-                ) : (
-                    <span
-                        className="w-100 dc__ellipsis-right"
-                        dangerouslySetInnerHTML={{
-                            __html: option.label.replace(
-                                new RegExp(formatOptionLabelMeta.inputValue, 'gi'),
-                                (highlighted) => `<mark>${highlighted}</mark>`,
-                            ),
-                        }}
-                    />
-                )}
-            </div>
-        )
-    }
+    const formatOptionLabel = useCallback(
+        (option: K8sObjectOptionType, formatOptionLabelMeta: FormatOptionLabelMeta<K8sObjectOptionType>) => {
+            return (
+                <div className="flex left column">
+                    {!formatOptionLabelMeta.inputValue ? (
+                        <span className="w-100 dc__ellipsis-right">{option.label}</span>
+                    ) : (
+                        <span
+                            className="w-100 dc__ellipsis-right"
+                            dangerouslySetInnerHTML={{
+                                __html: option.label.replace(
+                                    new RegExp(formatOptionLabelMeta.inputValue, 'gi'),
+                                    (highlighted) => `<mark>${highlighted}</mark>`,
+                                ),
+                            }}
+                        />
+                    )}
+                </div>
+            )
+        },
+        [],
+    )
 
-    function customFilter(option, searchText) {
-        const lowerLabel = option.data.label.toLowerCase()
+    const getOptionLabel = (option: K8sObjectOptionType) => {
+        const lowerLabel = option.label.toLowerCase()
         const lowerSearchText = searchText.toLowerCase()
-        return lowerLabel === k8Abbreviates?.[lowerSearchText] || lowerLabel.includes(lowerSearchText)
+        const abbreviateExpanded = k8Abbreviates?.[lowerSearchText]
+        return abbreviateExpanded === lowerLabel ? lowerSearchText : lowerLabel
     }
 
     const noOptionsMessage = () => 'No matching kind'
@@ -342,9 +271,10 @@ const Sidebar = ({
                 <ReactSelect
                     ref={searchInputRef}
                     placeholder="Jump to Kind"
-                    options={k8sObjectOptionsList}
+                    options={bringMatchedAbbreviatedOptionToFront(k8sObjectOptionsList)}
                     value={k8sObjectOptionsList[0]} // Just to enable clear indicator
                     inputValue={searchText}
+                    getOptionValue={getOptionLabel}
                     onInputChange={handleInputChange}
                     onChange={handleOnChange}
                     onBlur={hideMenu}
@@ -355,7 +285,6 @@ const Sidebar = ({
                     isSearchable
                     isClearable
                     formatOptionLabel={formatOptionLabel}
-                    filterOption={customFilter}
                     noOptionsMessage={noOptionsMessage}
                     classNamePrefix="kind-search-select"
                     styles={KIND_SEARCH_COMMON_STYLES}
@@ -436,7 +365,7 @@ const Sidebar = ({
                                 >
                                     <DropDown
                                         className={`${k8sObject.isExpanded ? 'fcn-9' : 'fcn-5'} rotate icon-dim-24 pointer`}
-                                        style={{ ['--rotateBy' as any]: !k8sObject.isExpanded ? '-90deg' : '0deg' }}
+                                        style={{ ['--rotateBy' as string]: !k8sObject.isExpanded ? '-90deg' : '0deg' }}
                                     />
                                     <span
                                         className="fs-14 fw-6 pointer w-100 pt-6 pb-6"
